go/bundle: Move decoder / encoder code to separate files
